refactor(core): extract update()'s inline commit into commitRouteUpdate (#1049)
The four route-CRUD commit cores were inconsistently located: add/replace
(adoptRouteArtifacts), remove (commitTreeChanges), clear (resetStore) are named
functions in routesStore.ts, but update()'s field-patch PREPARE/COMMIT was
hand-orchestrated inline in getRoutesApi.ts. Move it — and its four update-only
helpers (prepareForwardTo / prepareCustomFields / commitScalarConfig /
commitGuardUpdate, which must move down a layer to avoid a circular import) —
into a named commitRouteUpdate() in routesStore.ts beside its three siblings.

Pure co-location/naming (NOT the unifying primitive rejected in #1036 — four
distinct cores remain; update stays NO_TREE_REBUILD). Atomicity (#951), the
#956 seam, #952, #797 custom-fields, #967 async-forwardTo, and the #1046
handler-limit pre-flight all preserved. commitRouteUpdate returns the structural
fields for the caller's TREE_CHANGED emit so core reads each user getter once.
update() now reads symmetrically with its siblings. No behavior/public-API change.
